requirements.txt
gradio>=4.0
rembg[cpu]
opencv-python-headless
Pillow
shapely
trimesh[easy]
numpy
